mysite/requirements.txt
django==4.2.0
djangorestframework==3.14.0
django-cors-headers==4.3.1

rasa==3.6.0
rasa-sdk==3.6.0
sanic-cors==2.0.1

sqlalchemy==1.4.54
requests==2.31.0
python-dateutil==2.8.2
pillow==10.2.0

boto3
awscli
python-decouple==3.8
openai
